🚧 Last try with vico
It showed the points using a LinearChart with points
and no lines. But it crashed whenever any x values
(distance) matched.
This is because the linear plot is suposed to
draw functions and some kind of progress on the x.
Because drawing gas station at the same distance is
something we find very often, a bug repport has
been filled https://github.com/patrykandpatrick/vico/issues/1230
and now we will try a different library.